Inclusión boton tomar foto
