+++ app/src/main/AndroidManifest.xml
@@ -23,6 +23,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".Home"
+            android:label="@string/mi_activity_home" />
     </application>
 
 </manifest>
