+++ app/src/main/java/com/example/bachefinder/Home.java
@@ -1,0 +1,75 @@
+package com.example.bachefinder;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.TextView;
+import android.widget.Toast;
+import androidx.appcompat.app.AppCompatActivity;
+import java.io.IOException;
+import okhttp3.Call;
+import okhttp3.Callback;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
+
+public class Home extends AppCompatActivity {
+    public TextView textResponse;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.home);
+
+        Button takePhoto = findViewById(R.id.btn_take_photo);
+        takePhoto.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Toast.makeText(Home.this, "Botón capturar", Toast.LENGTH_SHORT).show();
+            }
+        });
+
+        Button validatePhoto = findViewById(R.id.btn_validate_photo);
+        validatePhoto.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                sendRequest();
+            }
+        });
+
+        textResponse = findViewById(R.id.text_validate_photo);
+    }
+
+    private void sendRequest() {
+        OkHttpClient client = new OkHttpClient();
+
+        String url = "http://192.168.100.15:8000/";
+
+        Request request = new Request.Builder()
+                .url(url)
+                .build();
+
+        client.newCall(request).enqueue(new Callback() {
+            @Override
+            public void onFailure(Call call, IOException e) {
+                runOnUiThread(new Runnable() {
+                    @Override
+                    public void run() {
+                        textResponse.setText("No fue posible conectar con el backend");
+                    }
+                });
+            }
+
+            @Override
+            public void onResponse(Call call, Response response) throws IOException {
+                String responseBody = response.body().string();
+                runOnUiThread(new Runnable() {
+                    @Override
+                    public void run() {
+                        textResponse.setText(responseBody);
+                    }
+                });
+            }
+        });
+    }
+}
