+++ app/src/main/java/com/example/bachefinder/MainActivity.java
@@ -1,53 +1,19 @@
 package com.example.bachefinder;
 
+import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.Toast;
+import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
-import java.io.IOException;
-import okhttp3.Call;
-import okhttp3.Callback;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
 
 public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        EdgeToEdge.enable(this);
         setContentView(R.layout.activity_main);
 
-        Button myButton = findViewById(R.id.btn_genearet); // Inicializa el botón
-        myButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                sendRequest();
-                Toast.makeText(MainActivity.this, "Botón presionado", Toast.LENGTH_SHORT).show();
-            }
-        });
-    }
-
-    private void sendRequest() {
-        OkHttpClient client = new OkHttpClient();
-
-        String url = "http://192.168.100.15:8000/";
-
-        Request request = new Request.Builder()
-                .url(url)
-                .build();
-
-        client.newCall(request).enqueue(new Callback() {
-            @Override
-            public void onFailure(Call call, IOException e) {
-                System.out.println("No conecta con el backend");
-            }
-
-            @Override
-            public void onResponse(Call call, Response response) throws IOException {
-                System.out.println(response.body().string());
-            }
-        });
+        Intent intent = new Intent(this, Home.class);
+        startActivity(intent);
     }
 }
