+++ app/src/main/res/values-night/themes.xml
@@ -2,6 +2,8 @@
     <!-- Base application theme. -->
     <style name="Base.Theme.BacheFinder" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your dark theme here. -->
+        <item name="android:background">@color/white</item>
+        <item name="android:textColor">@color/black</item>
         <!-- <item name="colorPrimary">@color/my_dark_primary</item> -->
     </style>
 </resources>
